Refactor project structure: migrate to TypeScript, update dependencies, and implement routing
