+++ src/App.tsx
@@ -1,6 +1,6 @@
 import { BrowserRouter as Router, Routes, Route, Link } from "react-router-dom";
+import Home from "@/pages/Home";
 import About from "@/pages/About";
-import Home from "@/pages/home";
 
 export default function App() {
     return (
