+++ tsconfig.node.json
@@ -1,0 +1,8 @@
+{
+    "compilerOptions": {
+        "composite": true,
+        "module": "ESNext",
+        "moduleResolution": "Node"
+    },
+    "include": ["vite.config.ts"]
+}
